xmonad: position dialogs above current window, not below
For tiled windows "below" means lower in the order of windows
displayed, which I like better than the default as it means the master
window stays the master when you open a new window from it.

For floating dialogs, however, "below" means *underneath* the
currently focussed window if the current is also floating. This meant
that dialogs which opened further dialogs would have the child dialogs
invisible; extremely irritating as they are usually modal and thus you
can't interact with the visible parent dialog.