remove wrong gtk3 dep for xmonad-custom
Git blame suggests I thought I needed the system package gtk3 to be in
the build environment for something to do with the wrapGAppsHook.

However the way I had actually done it would have been pulling in
the *Haskell* package gtk3, not the system gtk3 library. It's unlikely
this was doing anything at all; although it's remotely possible it was
pulling in some *propagated* dependency via indirectly depending on
gtk3. Removing it doesn't seem to hurt now, however.