seems to resolve issue with second monitor missing layout commands
I think this is just that moving the rivertile command to
river.settings.spawn rather than putting it in extraConfig has moved
it slightly earlier in the generated config, so the layout is running
by the time the layout commands to switch the main location are run. It
doesn't really resolve the race condition.